refactor: extract AddressFields to new component
